requirements.txt
flask
python-dotenv
replit-object-storage
werkzeug
gunicorn
psycopg2-binary
